Add CustomEmbeddingFunction class to mutualize code
